requirements.txt
regex
tqdm
diffusers
transformers
scipy
accelerate
opencv-python
Xformers
gradio
torch
torchvision
Pillow
matplotlib
numpy
git+https://github.com/facebookresearch/segment-anything.git
pycocotools 
onnxruntime 
onnx
